agentwithui/pyproject.toml
[project]
name = "agentwithui"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.8.3",
    "dotenv>=0.9.9",
    "openai-agents>=0.4.0",
    "streamlit>=1.50.0",
]
